Remove old example tests and add new tests for weewx_ha
